goals and users synced together

+++ backend/controllers/goalControllers.js
@@ -1,11 +1,12 @@
 const asyncHandler = require('express-async-handler')
 const Goal = require('../models/goalModel')
+const User = require('../models/userModel')
 
 // @desc    Get Goals
 // @route   GET /api/goals
 // @access  Private
 const getGoals = asyncHandler(async (req, res) => {
-    const goals = await Goal.find()
+    const goals = await Goal.find({ user: req.user.id })
 
     res.status(200).json(goals);
 })
@@ -20,6 +21,7 @@ const setGoal = asyncHandler(async (req, res) => {
     }
 
     const newGoal = await Goal.create({
+        user: req.user.id,
         text: req.body.text
     })
 
@@ -38,6 +40,17 @@ const updateGoal = asyncHandler(async (req, res) => {
         throw new Error('Goal not found!')
     }
 
+    const user = await User.findById(req.user.id)
+    if (!user) {
+        res.status(401)
+        throw new Error('User not found')
+    }
+
+    if (goal.user.toString() !== user.id) {
+        res.status(401)
+        throw new Error('User not authorized')
+    }
+
     const updatedGoal = await Goal.findByIdAndUpdate(id, req.body, { new: true })
 
     res.status(200).json(updatedGoal);
@@ -53,6 +66,17 @@ const deleteGoal = asyncHandler(async (req, res) => {
     if (!goal) {
         res.status(400)
         throw new Error('Goal not found!')
+    }
+
+    const user = await User.findById(req.user.id)
+    if (!user) {
+        res.status(401)
+        throw new Error('User not found')
+    }
+
+    if (goal.user.toString() !== user.id) {
+        res.status(401)
+        throw new Error('User not authorized')
     }
 
     await Goal.findByIdAndDelete(id);
